--- Exercises/Results/guzm3915/CS350 Exam (KC Guzman)/Social Media Exam/Social Media Exam.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Social Media Exam", "Social Media Exam\Social Media Exam.csproj", "{E851490C-916A-4D04-ADF2-9ADA95425D7D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E851490C-916A-4D04-ADF2-9ADA95425D7D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E851490C-916A-4D04-ADF2-9ADA95425D7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E851490C-916A-4D04-ADF2-9ADA95425D7D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E851490C-916A-4D04-ADF2-9ADA95425D7D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {19B8C558-500F-4632-B374-7D86F80EDBAA}
-	EndGlobalSection
-EndGlobal
