--- Exercises/Results/guzm3915/CS350 Exam (KC Guzman)/Social Media Exam/Social Media Exam/data.csv
@@ -1,1 +1,0 @@
-﻿userID,password,postID,friends
