+++ Exercises/Results/guzm3915/Exam3/Documentation.md
@@ -1,0 +1,26 @@
+# Documentation Markdown
+
+SystemFunc:
+
+ * resetdata rests user data
+ * exportdata exports the user data into data.csv
+ * importdata imports from data.csv
+
+Post:
+
+ * listtopic lists all topics fromt he user
+ * addtopic adds topics the user inputs
+ * removetopic removes the topic
+ 
+Friend
+
+ * viewFriend shows all friends to the user
+ * newFriendship adds friend based on input
+ * removeFriendship removes friendship based on input
+
+Person
+
+ * addFriend adds friends to the friends list of a user
+ * remove friend removes that friend from the list of a user
+ * addpost adds post and post ID to a post from the user
+ * remove post removes the post and post ID from a user
